Moved constant USD_CAMERA to usd_collection module.
